+++ src/com/nsu/rds/data/repositories/CourseRepository.java
@@ -38,7 +38,7 @@ public class CourseRepository {
             File myObj = new File(Const.ALL_COURSE_LIST);
             Scanner myReader = new Scanner(myObj);
             while (myReader.hasNext()) {
-                newList.add(new Courses(myReader.next(), myReader.next(), myReader.nextInt()));
+                newList.add(new Courses(myReader.next(), myReader.next(), myReader.nextDouble()));
             }
             myReader.close();
         } catch (FileNotFoundException e) {
